requirements.txt
Django==3.2.4
django-tinymce==3.3.0
graphene_django==2.15.0
python-dotenv==0.18.0